feat: remove mobile nav overlay
