+++ index.html
@@ -22,8 +22,7 @@
   <div id="notifications" role="alert" aria-live="polite"></div>
 
   <!-- Navigation -->
-  <button id="navToggle" class="nav-toggle btn btn-primary" aria-expanded="false" aria-controls="mainNav" aria-label="Menu"><span class="material-symbols-outlined nav-icon" aria-hidden="true">menu</span><span class="nav-text">Menu</span></button>
-  <nav id="mainNav" aria-hidden="true">
+  <nav id="mainNav">
     <ul class="nav-container">
       <li><a href="#" id="navCheckout" data-section="checkout" class="btn btn-primary"><span class="material-symbols-outlined nav-icon" aria-hidden="true">barcode_reader</span> Check-Out</a></li>
       <li><a href="#" id="navAdmin" data-section="admin" class="btn btn-primary"><span class="material-symbols-outlined nav-icon" aria-hidden="true">admin_panel_settings</span> Admin</a></li>
