+++ styles/main.css
@@ -228,17 +228,6 @@
       text-decoration: none;
     }
 
-      .nav-toggle {
-        display: none;
-        align-items: center;
-        gap: 0.25rem;
-        font-weight: bold;
-        padding: 0.5rem 0.9375rem;
-        width: auto;
-        flex: 0 1 auto;
-        margin: 0;
-    }
-
     .nav-icon {
       font-size: 1.2rem;
     }
@@ -346,65 +335,6 @@
     }
 
     @media (max-width: 37.5rem) {
-      nav {
-        display: none;
-        position: fixed;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100vh;
-        overflow-y: auto;
-        background-color: rgba(0,0,0,0.5);
-        z-index: 1000;
-        padding: 1.25rem;
-        box-sizing: border-box;
-        justify-content: center;
-        align-items: flex-start;
-      }
-
-      nav.show {
-        display: flex;
-      }
-
-      nav .nav-container {
-        background: #fff;
-        max-width: 37.5rem;
-        width: 100%;
-        margin: 1.25rem auto;
-        padding: 1.25rem;
-        border-radius: 0.5rem;
-        box-shadow: 0 0 0.625rem rgba(0,0,0,0.1);
-        display: flex;
-        flex-direction: column;
-        align-items: flex-start;
-        gap: 0.625rem;
-      }
-
-      nav .nav-container li {
-        flex: 1 1 auto;
-        width: auto;
-      }
-
-      nav .nav-container a {
-        padding: 0.375rem 0.75rem;
-      }
-
-      .nav-toggle {
-        display: flex;
-        position: fixed;
-        top: 0.5rem;
-        left: 0.5rem;
-        z-index: 1001;
-      }
-
-      .nav-toggle.open {
-        left: 0.5rem;
-      }
-
-      .nav-icon {
-        font-size: 1rem;
-      }
-
       .employeeRow,
       .equipmentRow {
         flex-direction: column;
@@ -430,11 +360,11 @@
       }
 
       .records-container,
-    form {
+      form {
         max-width: 100%;
         margin: 0.625rem auto;
         padding: 0.625rem;
-    }
+      }
     }
 
   .list-container {
